# doc/api.yaml
openapi: 3.0.0
info:
  description: |-
    This concerns the "Web and Software Architecture" course to develop from scratch a web application that allows users to share photos of special moments.
    It allows users to like, comment on posts, ban other users, and so forth.

    Note: "Banning" and "setting a profile to private" are different. Setting a profile to public (private = false), means that any user except those whom he bans can see the user's profile, comment, like, etc without following the user. This is a feature much like "explore" in many social media. If a user wants only his followers to see his feeds except those banned, he should set it to private. Otherwise, even those who did not follow him and were not banned by him can see his feeds.
  title: Photo Sharing App
  version: 1.0.0
  contact:
    name: Yusupha Juwara
    url: https://github.com/YusuphaJuwara
    email: yusuphajuwara@gmail.com
servers:
  - url: http://localhost:3000
    description: Local Development Server
components:
  parameters:
...
    userID:  #Arbitrary name for the definition that will be used to refer to it. Not necessarily the same as the parameter name.
      in: path
      name: user-id
      required: true
      schema:   
        $ref: "#/components/schemas/ID"
      description: The user ID
    postID:
      in: path
      name: post-id
      required: true
      schema:   
        $ref: "#/components/schemas/ID"
      description: The post ID
...
    followID:
      in: path
      name: follow-id
      required: true
      schema:   
        $ref: "#/components/schemas/ID"
      description: A user ID
...
    hashtag:
      in: path
      name: hashtag
      required: true
      schema:   
        $ref: "#/components/schemas/hashtag"
      description: A hashtag ID
...
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      bearerFormat: uuid    # optional, arbitrary value for documentation purposes
  responses:
    NoContent204:
      description: Successful.
    BadRequest400:
      description: |-
        The request was not compliant with the documentation (eg. missing fields, etc).
        Further info in server logs.
    UnAuth401:
      description: |-
        Access token is missing or invalid.
        The request was not authorized. The user must be authorized first to perform this request.
    Forbidden403:
      description: |-
        Trying to access a resource that the user should not. This is forbidden action.
        E.g., 
        a banned user trying to get an info about the user who banned him, 
        or trying to delete a comment when the user is not the one who placed the comment or on whose post the comment was placed, etc.
        Further info in server logs.
    NotFound404:
      description: Not found.
    InternalServerError500:
      description: |-
        The server encountered an internal error.
        Further info in server logs.
  schemas:
    ID:
      description: The unique identifier. It is 36 characters containing 4 "-". 
      type: string
      format: uuid

      pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$'
      minLength: 36
      maxLength: 36
      example: 123e4567-e89b-12d3-a456-426655440000
...
    hashtag:
      description: |-
        Hashtag contains at least one alphabet and zero or more other alphanumeric characters in [a-zA-Z0-9].
      type: string
      pattern: '^(?=.*?[a-zA-Z])[a-zA-Z0-9]{1,20}$'
      minLength: 1
      maxLength: 20
      example: "covid19"
    hashtags:
      description: Every post also includes the hashtags
      type: array
      minItems: 0
      maxItems: 100
      items:  
        $ref: '#/components/schemas/hashtag'
...
    ids-array:
      description: array of uuid's.
      type: array
      minItems: 0
      maxItems: 1000
      items:  
        $ref: '#/components/schemas/ID'
...
security:
  - bearerToken: []
...
paths:
...
  /users/{user-id}/follow:
    parameters:
      - $ref: "#/components/parameters/userID"
    get:
      tags: [Follow]
      operationId: getUserFollows
      summary: Get the user IDs of the users whom the user follows or who follows the user.
      description: |-
        Returns an object which contains 2 arrays of user IDs; one for whom the user follows, and the other for who follows the user.
      responses:
        "200":
          description: An object containing 2 arrays, one for the user's followers and the other for the user's followings.
          content:
            application/json:
              schema:
                description: "An object containing 2 arrays."
                type: object
                properties:
                  followers-array:  
                    $ref: '#/components/schemas/ids-array'
                  followings-array:   
                    $ref: '#/components/schemas/ids-array'
        "400":  
          $ref: "#/components/responses/BadRequest400"
        "401":  
          $ref: "#/components/responses/UnAuth401"
        # "403":  
        #   $ref: "#/components/responses/Forbidden403"
        "404":  
          $ref: "#/components/responses/NotFound404"
        "500":  
          $ref: "#/components/responses/InternalServerError500"
  /users/{user-id}/follow/{follow-id}:
    parameters:
      - $ref: "#/components/parameters/userID"
      - $ref: "#/components/parameters/followID"
    put:
      tags: [Follow]
      operationId: followUser
      summary: A user follows another user.
      description: |-
        A user with the user-id follows another user with follow-id, provided that he didn't already.
        When a user bans another user, the banned user cannot follow the other user. In this case a forbidden error is returned
      responses:
        "204":  
          $ref: "#/components/responses/NoContent204"
        "400":  
          $ref: "#/components/responses/BadRequest400"
        "401":  
          $ref: "#/components/responses/UnAuth401"
        "403":  
          $ref: "#/components/responses/Forbidden403"
        "404":  
          $ref: "#/components/responses/NotFound404"
        "500":  
          $ref: "#/components/responses/InternalServerError500"
    delete:
      tags: [Follow]
      operationId: unfollowUser
      summary: a user unfollows another user.
      description: |-
        A user with the user-id unfollows another user with follower-id.
      responses:
        "204":  
          $ref: "#/components/responses/NoContent204"
        "400":  
          $ref: "#/components/responses/BadRequest400"
        "401":  
          $ref: "#/components/responses/UnAuth401"
        "403":  
          $ref: "#/components/responses/Forbidden403"
        "404":  
          $ref: "#/components/responses/NotFound404"
        "500":  
          $ref: "#/components/responses/InternalServerError500"
...
  /users/{user-id}/posts/{post-id}/hashtags/{hashtag}:
    parameters:
      - $ref: "#/components/parameters/userID"
      - $ref: "#/components/parameters/postID"
      - $ref: '#/components/parameters/hashtag'
    put:
      tags: [Hashtag]
      operationId: addHashtag
      summary: Add a hashtag to a post.
      description: |-
        This adds a hashtag on a post.
        Hashtag can be added to any post through this endpoint after the post has been posted, or by adding the hashtag directly when posting the posts.
      responses:
        "201":
          description: Hashtag created successfully
          content:
            application/json:
              schema:   
                $ref: "#/components/schemas/hashtag"
        "204":  
          $ref: "#/components/responses/NoContent204"
        "400":  
          $ref: "#/components/responses/BadRequest400"
        "401":  
          $ref: "#/components/responses/UnAuth401"
        "403":  
          $ref: "#/components/responses/Forbidden403"
        # "404":  
        #   $ref: "#/components/responses/NotFound404"
        "500":  
          $ref: "#/components/responses/InternalServerError500" 
    delete:
      tags: [Hashtag]
      operationId: deleteHashtag
      summary: Deletes a hashtag on a given post.
      description: |-
        This simply deletes the specified hashtag.
      responses:
        "204":  
          $ref: "#/components/responses/NoContent204"
        "400":  
          $ref: "#/components/responses/BadRequest400"
        "401":  
          $ref: "#/components/responses/UnAuth401"
        "403":  
          $ref: "#/components/responses/Forbidden403"
        # "404":  
        #   $ref: "#/components/responses/NotFound404"
        "500":  
          $ref: "#/components/responses/InternalServerError500"
  /posts/{post-id}/hashtags/:
    parameters:
      - $ref: "#/components/parameters/postID"
    get:
      tags: [Hashtag]
      operationId: getPostHashtags
      summary: An array of hashtags of a post.
      description: |-
        This returns an array of hashtags of a given post. That is, an array of strings.
      responses:
        "200":
          description: Returns an array of hashtags of a given post
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/hashtags' 
        "400":  
          $ref: "#/components/responses/BadRequest400"
        "401":  
          $ref: "#/components/responses/UnAuth401"
        # "403":  
        #   $ref: "#/components/responses/Forbidden403"
        "404":  
          $ref: "#/components/responses/NotFound404"
        "500":  
          $ref: "#/components/responses/InternalServerError500" 
...